pyproject.toml
[tool.poetry]
name = "ev-charger-ap"
version = "0.1.0"
description = ""
authors = ["jasper200207"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.9.0,<3.9.1 || >3.9.1,<4.0"
environs = "^14.1.0"
cryptography = "44.0.0"
pydantic = "1.10.21"
psutil = "^6.1.1"
py4j = "^0.10.9"
types-toml = "^0.10.8.7"
mypy-extensions = "^1.0.0"
types-psutil = "^6.1.0.20241221"
kivy = {extras = ["base"], version = "^2.3.1"}
kivymd2 = "^2.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = ">=0.18.3,<0.26.0"
asynctest = "^0.13.0"
freezegun = "^1.2.1"
black = "^25.1.0"
isort = "^6.0.0"
flake8 = ">=4.0.1,<7.2.0"
pytest-cov = ">=3,<7"
aiofile = "^3.7.4"
toml = "^0.10.2"
mypy = "^1.14.1"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "strict"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
